Fantasy-Scoring auf Sportmonks, Vereinsseite, Aufstellung & UI-Vereinheitlichung
Version 1.1.0+3.

Fantasy-Stats
- sync-stats zieht den vollen Roh-Stat-Satz aus Sportmonks statt Tore/Zu-Null
  aus OpenLigaDB. Fixture-IDs kommen aus public.fixtures (kostet keinen
  API-Request), /fixtures/multi liefert bis zu 25 Spiele für einen Request.
  Zuordnung über sportmonks:<id> — kein Raten über Nachnamen mehr.
- Migration 0074: ~21 Spalten für die Zählwerte aus PlayerEvents, dazu `source`,
  damit ein OpenLigaDB-Notfalleintrag nicht wie ein vollständiger Satz aussieht.
- Dart-Scoring auf die verbindliche Punktevergabe gehoben (Einsatzstufen,
  Offensive, Zu Null ab 60 Min, Paraden, Zweikämpfe, Meilensteine, Rating).
  Punkte sind dadurch double statt int — zieht sich durch Tabelle, Playoffs,
  Recap und die Fantasy-Oberfläche.
- Gehaltener Elfmeter: Sportmonks kennt den Typ nicht; nach Produktentscheidung
  bekommt der gegnerische Torwart die Punkte bei jedem missed_penalty.

Vereinsseite
- Neuer ClubScreen (Spielplan, Tabelle, Kader, News), erreichbar über jedes
  Vereinswappen. Kader über den neuen squad-Endpunkt (eigene Rate-Limit-Entity).
- Spielplan teilt sich die Darstellung mit dem Favoriten-Tab
  (widgets/team_fixture_list.dart) statt sie zu kopieren.

Live-Tabelle
- mergeLiveResults trägt laufende und frisch beendete Spiele nach, die
  Sportmonks noch nicht fortgeschrieben hat. Die API-Tabelle bleibt Grundlage,
  weil nur sie Punktabzüge kennt.

Spieldetails
- Heim oben, Auswärts unten; Trikot-Symbole in Vereinsfarben statt Kreisen,
  Formation in der Seitenbeschriftung im Feld.
- Bank mit Mannschaftszuordnung; Wechsel-Marken am Spieler, Eingewechselte
  bleiben auf der Bank mit "für <Name>".

Oberfläche
- Navi-Leiste flacher und ohne gefüllte Auswahl-Pille.
- Live-Tab: eine Box je Liga, darin nach Anstoßzeit, Zeitblöcke fein getrennt.
- Alle acht Tab-Leisten auf SegmentedTabBar; Transfer-Umschalter und
  Filter-Pille auf PillSelector (vorher Material-Vorgabe secondaryContainer).

Anmeldung
- currentUserProvider liest den Nutzer aus dem Auth-Ereignis statt aus dem
  globalen Client — sonst blieb das Gate nach erfolgreichem Login stehen.
- Passwort-Reset greift jetzt auch beim Warmstart: main() lief dabei nicht
  erneut, der Link wurde also nie gelesen.

Werkzeug
- Golden-Vorschauen für Trikots, Aufstellung, Tab-Leiste und Pillen — die
  synthetischen Simulator-Taps sind unzuverlässig, die Goldens haben mehrere
  Fehler aufgedeckt (u. a. vertauschte Spielfeldhälften).

+++ lib/features/fantasy/logic/bracket.dart
@@ -55,8 +55,8 @@ class BracketMatch {
 
   final BracketSlot home;
   final BracketSlot away;
-  final int homePoints;
-  final int awayPoints;
+  final double homePoints;
+  final double awayPoints;
 
   /// Alle Spieltage der Partie sind beendet → Ergebnis steht.
   final bool decided;
@@ -131,16 +131,16 @@ PlayoffBracket buildPlayoffBracket({
   required int playoffTeams,
   required int startRound,
   required int weeksPerRound,
-  required Map<int, Map<String, int>> roundTotals,
+  required Map<int, Map<String, double>> roundTotals,
   required Set<int> finishedMatchdays,
 }) {
   final n = seeding.length;
   final pt = playoffTeams.clamp(0, n);
 
   // Punkte eines Teams in Playoff-Runde [r] (Summe über die Runden-Spieltage).
-  int roundPoints(String id, int r) {
+  double roundPoints(String id, int r) {
     final md0 = startRound + r * weeksPerRound;
-    var sum = 0;
+    var sum = 0.0;
     for (var w = 0; w < weeksPerRound; w++) {
       sum += roundTotals[md0 + w]?[id] ?? 0;
     }
