+++ lib/features/fantasy/ui/fantasy_table_screen.dart
@@ -80,7 +80,7 @@ class FantasyTableBody extends ConsumerWidget {
         return pa != pb ? pa.compareTo(pb) : a.compareTo(b);
       });
 
-    final totalsByRound = <int, Map<String, int>>{
+    final totalsByRound = <int, Map<String, double>>{
       for (final entry in seasonStats.entries)
         entry.key: effectiveTotalsForRound(
           stats: entry.value,
